Summary code cleanup
I can't believe I had an unless/else in this code, for serious.
And what's the pointless format_time method in the spec all about?
I'm a freaking moron.

Also, a problem with OS X's strftime gives some problems with summary
of a punch-in project during part of the year. Not sure what to do about that.