fix YAML processing errors in test
Apparently this is a problem with how YAML parsing has changed over the
years. And `String#strip_heredoc` doesn't exist in core Ruby, which is
what I'd like to be using. I could go for "squiggly heredoc", but that
requires Ruby 2.3, and I don't need to do that. This is simple enough.